fix of file loading

+++ src/frontend/util.py
@@ -1,5 +1,6 @@
 from frontend.setup import frontend
 import urllib
+import os
 
 def user_picture(user_id):
     # Смотрим случай, когда есть аватарка
@@ -11,9 +12,10 @@ def user_picture(user_id):
         file_id = pictureMas[-1].file_id
     # Случай, когда нет аватарки (мб выдавать месс, мол, поставьте аву, пж?)
     except IndexError:
-        url = 'https://i.ibb.co/m5NDJqv/noavatar.jpg'
-        f = open('out.jpg', 'wb')
-        f.write(urllib.request.urlopen(url).read())
+        if not os.path.isfile('out.jpg'):
+            url = 'https://i.ibb.co/m5NDJqv/noavatar.jpg'
+            f = open('out.jpg', 'wb')
+            f.write(urllib.request.urlopen(url).read())
         file_id = open('out.jpg', 'rb')
 
     return file_id
